+++ core/services.go
@@ -1067,6 +1067,28 @@ func (app *Application) syncAuthmanGroupMemberships(clientID string, authmanGrou
 	return nil
 }
 
+func (app *Application) sendGroupNotification(clientID string, notification model.GroupNotification) error {
+	memberStatuses := notification.MemberStatuses
+	if len(memberStatuses) == 0 {
+		memberStatuses = []string{"admin", "member"}
+	}
+
+	members, err := app.findGroupMemberships(clientID, model.MembershipFilter{
+		GroupIDs: []string{notification.GroupID},
+		UserIDs:  notification.MemberIDs.ToUserIDs(),
+	})
+
+	if err != nil {
+		return err
+	}
+
+	app.sendNotification(members.GetMembersAsNotificationRecipients(func(member model.GroupMembership) bool {
+		return true
+	}), notification.Topic, notification.Subject, notification.Body, notification.Data)
+
+	return nil
+}
+
 func (app *Application) sendNotification(recipients []notifications.Recipient, topic *string, title string, text string, data map[string]string) {
 	app.notifications.SendNotification(recipients, topic, title, text, data)
 }