+++ docs/docs.go
@@ -2015,6 +2015,59 @@ const docTemplate = `{
                 }
             }
         },
+        "/api/int/group/{group-id}/notification": {
+            "post": {
+                "security": [
+                    {
+                        "IntAPIKeyAuth": []
+                    }
+                ],
+                "description": "Sends a notification to members of a group",
+                "consumes": [
+                    "application/json"
+                ],
+                "tags": [
+                    "Internal"
+                ],
+                "operationId": "SendGroupNotification",
+                "parameters": [
+                    {
+                        "type": "string",
+                        "description": "APP",
+                        "name": "APP",
+                        "in": "header",
+                        "required": true
+                    },
+                    {
+                        "type": "string",
+                        "description": "Run this request asynchronously. Possible values: 'true' will return with status code 200 no matter of producing an error.  or 'false' - it will wait until the request finished the job and return with the real status code and/or error. Default: 'false'",
+                        "name": "async",
+                        "in": "query"
+                    },
+                    {
+                        "description": "body data",
+                        "name": "data",
+                        "in": "body",
+                        "required": true,
+                        "schema": {
+                            "$ref": "#/definitions/sendGroupNotificationRequestBody"
+                        }
+                    },
+                    {
+                        "type": "string",
+                        "description": "Group ID",
+                        "name": "group-id",
+                        "in": "path",
+                        "required": true
+                    }
+                ],
+                "responses": {
+                    "200": {
+                        "description": ""
+                    }
+                }
+            }
+        },
         "/api/int/group/{identifier}": {
             "get": {
                 "security": [
@@ -2984,6 +3037,14 @@ const docTemplate = `{
                 }
             }
         },
+        "MemberRef": {
+            "type": "object",
+            "properties": {
+                "user_id": {
+                    "type": "string"
+                }
+            }
+        },
         "MembershipFilter": {
             "type": "object",
             "properties": {
@@ -3564,6 +3625,43 @@ const docTemplate = `{
                     "type": "string"
                 },
                 "title": {
+                    "type": "string"
+                }
+            }
+        },
+        "sendGroupNotificationRequestBody": {
+            "type": "object",
+            "required": [
+                "body",
+                "subject"
+            ],
+            "properties": {
+                "body": {
+                    "type": "string"
+                },
+                "data": {
+                    "type": "object",
+                    "additionalProperties": {
+                        "type": "string"
+                    }
+                },
+                "member_statuses": {
+                    "description": "default: [\"admin\", \"member\"]",
+                    "type": "array",
+                    "items": {
+                        "type": "string"
+                    }
+                },
+                "members": {
+                    "type": "array",
+                    "items": {
+                        "$ref": "#/definitions/MemberRef"
+                    }
+                },
+                "subject": {
+                    "type": "string"
+                },
+                "topic": {
                     "type": "string"
                 }
             }