+++ driver/web/rest/internal.go
@@ -24,6 +24,7 @@ import (
 	"log"
 	"net/http"
 	"strconv"
+	"strings"
 
 	"github.com/gorilla/mux"
 )
@@ -410,5 +411,90 @@ func (h *InternalApisHandler) DeleteGroupEvent(clientID string, w http.ResponseW
 	}
 
 	w.Header().Set("Content-Type", "text/plain")
+	w.WriteHeader(http.StatusOK)
+}
+
+// sendGroupNotificationRequestBody Request body wrapper for sending a notifications to members of a group
+type sendGroupNotificationRequestBody struct {
+	MemberStatuses []string          `json:"member_statuses"` // default: ["admin", "member"]
+	MemberIDs      model.MemberIDs   `json:"members"`
+	Subject        string            `json:"subject" validate:"required"`
+	Topic          *string           `json:"topic"`
+	Body           string            `json:"body" validate:"required"`
+	Data           map[string]string `json:"data"`
+} // @name sendGroupNotificationRequestBody
+
+// SendGroupNotification Sends a notification to members of a group
+// @Description Sends a notification to members of a group
+// @ID SendGroupNotification
+// @Tags Internal
+// @Accept json
+// @Param APP header string true "APP"
+// @Param async query string false "Run this request asynchronously. Possible values: 'true' will return with status code 200 no matter of producing an error.  or 'false' - it will wait until the request finished the job and return with the real status code and/or error. Default: 'false'"
+// @Param data body sendGroupNotificationRequestBody true "body data"
+// @Param group-id path string true "Group ID"
+// @Success 200
+// @Security IntAPIKeyAuth
+// @Router /api/int/group/{group-id}/notification [post]
+func (h *InternalApisHandler) SendGroupNotification(clientID string, w http.ResponseWriter, r *http.Request) {
+	//validate input
+	params := mux.Vars(r)
+	groupID := params["group-id"]
+	if len(groupID) <= 0 {
+		log.Println("group-id is required")
+		http.Error(w, "group-id is required", http.StatusBadRequest)
+		return
+	}
+
+	asyncBoolValue := false
+	asyncValue := getStringQueryParam(r, "async")
+	if asyncValue != nil {
+		asyncBoolValue = strings.ToLower(*asyncValue) == "true"
+	}
+
+	data, err := ioutil.ReadAll(r.Body)
+	if err != nil {
+		log.Printf("Error on read the sendGroupNotificationRequestBody - %s\n", err.Error())
+		http.Error(w, http.StatusText(http.StatusBadRequest), http.StatusBadRequest)
+		return
+	}
+
+	var requestData sendGroupNotificationRequestBody
+	err = json.Unmarshal(data, &requestData)
+	if err != nil {
+		log.Printf("Error on unmarshal the sendGroupNotificationRequestBody - %s\n", err.Error())
+		http.Error(w, err.Error(), http.StatusBadRequest)
+		return
+	}
+
+	validate := validator.New()
+	err = validate.Struct(requestData)
+	if err != nil {
+		log.Printf("Error on validating sendGroupNotificationRequestBody - %s\n", err.Error())
+		http.Error(w, err.Error(), http.StatusBadRequest)
+		return
+	}
+
+	notification := model.GroupNotification{
+		GroupID:        groupID,
+		MemberIDs:      requestData.MemberIDs,
+		MemberStatuses: requestData.MemberStatuses,
+		Subject:        requestData.Subject,
+		Body:           requestData.Body,
+		Topic:          requestData.Topic,
+		Data:           requestData.Data,
+	}
+
+	if asyncBoolValue {
+		go h.app.Services.SendGroupNotification(clientID, notification)
+	} else {
+		err = h.app.Services.SendGroupNotification(clientID, notification)
+		if err != nil {
+			log.Println(err.Error())
+			http.Error(w, err.Error(), http.StatusInternalServerError)
+			return
+		}
+	}
+
 	w.WriteHeader(http.StatusOK)
 }